AprillTagDatabase class added and refactored code
